move parse_table_wcs into CubeCutout as a static method.
